kesho/securecourse-pro/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
slowapi==0.1.9
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
boto3==1.34.0
cloudflare==2.11.7